Added emergency contact for deletion

+++ Capstone_Project/Back-End/(Python)/routes/customer.py
@@ -151,6 +151,9 @@ def customer_remove():
     delete_waiver=sql_functions.execute_query(sql_connection,"delete from waiver where customer_id = %s",(customer_id,))
     if type(delete_waiver)==int:
         return make_response("Unable to delete from waiver",503)
+    delete_emergency=sql_functions.execute_query(sql_functions,"delete from emergency_contact where customer_id=%s",(customer_id,))
+    if type(delete_emergency)==int:
+        return make_response("Unable to delete emergency contact",503)
     delete_customer=sql_functions.execute_query(sql_connection,"delete from customer where customer_id=%s",(customer_id,))
     if type(delete_customer)==int:
         return make_response("Unable to delete customer",503)
